# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/accounts/{tenant}/reset/getResetPasswordToken": {
            "post": {
                "tags": [
                    "reset"
                ],
                "summary": "get reset password token",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/{tenant}/reset/resetPassword": {
            "post": {
                "tags": [
                    "reset"
                ],
                "summary": "reset password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "new password",
                        "name": "newPassword",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/{tenant}/reset/smsAvailable": {
            "get": {
                "tags": [
                    "reset"
                ],
                "summary": "check sms available",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/{tenant}/reset/verifyResetPasswordRequest": {
            "post": {
                "tags": [
                    "reset"
                ],
                "summary": "forgot password",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "verify method",
                        "name": "verifyMethod",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pass code payload",
                        "name": "passCodePayload",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "+86",
                        "description": "area code",
                        "name": "areaCode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}